improve testing infrastructure
