+++ RandBLAS/skge.hh
@@ -7,8 +7,6 @@
 #include "RandBLAS/dense.hh"
 #include "RandBLAS/sparse_skops.hh"
 
-#include <blas.hh>
-
 #include <iostream>
 #include <stdio.h>
 #include <stdexcept>
